chore: improve indicators page
